requirements.txt
pandas
numpy
spacy
nltk
scikit-learn
sentence-transformers
matplotlib
plotly
streamlit